% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lw.R
\name{fit_lw}
\alias{fit_lw}
\title{Fit length-weight model}
\usage{
fit_lw(len_obs, wgt_obs, a_strt, b_strt, logSigma = 0)
}
\arguments{
\item{len_obs}{The observed length of fish.}

\item{wgt_obs}{The observed weight of fish.}

\item{a_strt}{The starting value for the a parameter.}

\item{b_strt}{The starting value for the b parameter.}

\item{logSigma}{The starting value for the logSigma parameter.
If no value specified, then 0 will be used.}
}
\value{
A summary of the estimated parameters and plot
}
\description{
Fit length-weight model using maximum likelihood estimation.
}
\examples{
#library(TMB)
fit_lw(len_obs, wgt_obs, a_strt, b_strt, logSigma = 0)
}
