% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_vbgf.R
\name{fit_vbgf}
\alias{fit_vbgf}
\title{Fit von Bertalanffy Growth Function}
\usage{
fit_vbgf(len_obs, age_obs, linf_strt, k_strt, t0_start, logsigma = 0)
}
\arguments{
\item{len_obs}{The observed length of fish.}

\item{age_obs}{The observed ages of fish.}

\item{linf_strt}{The starting value for L-infinity (average length at maximum age).}

\item{k_strt}{The starting value for Kappa - growth coefficent.}

\item{t0_start}{The starting value for t0 (age at zero length).}

\item{logSigma}{The starting value for the logsigma parameter.
If no value specified, then 0 will be used.}
}
\value{
A summary of the estimated parameters and plot
}
\description{
Fit von Bertalanffy Growth Function to length at age data using maximum likelihood estimation.
}
\examples{
#library(TMB)
fit_vbgf(len_obs, age_obs, linf_strt, k_strt, t0_start, logsigma = 0)
}
