% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbrm_fl_age.R
\docType{data}
\name{sbrm_fl_age}
\alias{sbrm_fl_age}
\title{Stonebream Length and Age Data}
\format{
Data in tabular format.
}
\source{
(\href{https://saambr.org.za/oceanographic-research-institute-ori/}{ORI})
}
\usage{
data(sbrm_fl_age)
}
\description{
Stonebream \emph{Neoscorpis lithophilus} length and age data.
}
\examples{
library(jmsat)
data(sbrm_fl_age)
}
\references{
Data collected and owned by the Oceanographic Research Institute, Durban, South Africa.
}
\keyword{datasets}
