% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbrm_wgt_fl.R
\docType{data}
\name{sbrm_wgt_fl}
\alias{sbrm_wgt_fl}
\title{Stonebream Weight and Length Data}
\format{
Data in tabular format.
}
\source{
(\href{https://saambr.org.za/oceanographic-research-institute-ori/}{ORI})
}
\usage{
data(sbrm_wgt_fl)
}
\description{
Stonebream \emph{Neoscorpis lithophilus} weight and length data.
}
\examples{
library(jmsat)
data(sbrm_wgt_fl)
}
\references{
Data collected and owned by the Oceanographic Research Institute, Durban, South Africa.
}
\keyword{datasets}
